requirement.txt
tensorflow==2.1.0
numpy==1.19.1
pandas==1.1.1
biopython==1.79
